fix(refresh): el job llevaba 8 corridas muertas por dos parámetros sin tipo
`refresh-opportunities` fallaba en su PRIMERA consulta, antes de tocar Mercado
Público: 0 encontradas, 0 s de duración, y en `error_details` sólo
`{"fatal":"operator is not unique: unknown - unknown"}`.

La causa es el cupo reservado que introdujo 6211341:

    LIMIT GREATEST($1 - $3, 1)

Por el protocolo extendido los parámetros llegan como `unknown`, y Postgres no
tiene un '-' preferido entre dos `unknown`, así que aborta la consulta ENTERA.
Contraste dentro de la misma query: `($2 || ' days')::interval` SÍ funciona sin
cast, porque para '||' el motor prefiere la categoría string y resuelve a text.
Por eso el problema se veía arbitrario.

POR QUÉ NO LO VIO LA VALIDACIÓN DE 6211341

Ese commit dice "Validado en producción: el lote pasa de 80/0 a 48 urgentes + 32
adjudicaciones", y era cierto. La prueba se hizo con los números escritos a mano,
donde son literales tipados y la consulta corre perfecto. Falla sólo al pasar por
el driver, y `tsc` no mira dentro del string. Tres capas verdes sobre un job
muerto.

ALCANCE

8 corridas fallidas seguidas. Última útil 2026-08-12 13:48 UTC, primera rota
22:03 UTC del mismo día: ~30 h sin refrescar estados ni detectar adjudicaciones,
que es justo la capacidad que 6211341 venía a arreglar.

VERIFICACIÓN (contra producción, no razonada)

· Reproducción aislada vía `pg`: la forma vieja devuelve textual
  `operator is not unique: unknown - unknown`; la nueva devuelve cupo=48.
· `FIND_REFRESH_CANDIDATES` completa con [80, 75, 32]: 80 filas, 48 en prioridad
  2 y 32 en prioridad 3 — el reparto que el commit prometía.
· Corrida real post-deploy, disparada por el camino del cron:
  success, found=80 ok=80 fail=0, 187 s, 26 cambios de estado y 69 eventos
  emitidos a Licitus (el atraso de las 30 h caídas).

EL TEST

`opportunity-sql.test.ts` recorre todos los `SQL` y falla si un parámetro entra
sin cast a una operación aritmética. Es lo único que puede atrapar esta clase de
bug sin base de datos: el typecheck no ve dentro del string y psql no reproduce
el caso. Comprobado que el guard marca `$1 - $3` y acepta `$1::int - $3::int`.

OJO: el status verde de GitHub Actions no sirve para vigilar esto. Los 8 fallos
salieron con Actions en verde — ese step sólo confirma que el endpoint aceptó el
POST. Lo que hay que mirar es la fila de `sync_logs`.

+++ services/mercado-publico/src/modules/opportunities/infrastructure/opportunity-sql.ts
@@ -158,7 +158,13 @@ export const SQL = {
             AND o.closing_at BETWEEN NOW() AND NOW() + INTERVAL '72 hours'
         ) b ORDER BY id, priority
       ) d ORDER BY priority ASC, orden ASC NULLS LAST
-      LIMIT GREATEST($1 - $3, 1)
+      -- Los ::int NO son decorativos. Sin ellos los dos parámetros llegan sin tipo
+      -- por el protocolo extendido y Postgres aborta la consulta ENTERA con
+      -- "operator is not unique: unknown - unknown": no existe un '-' preferido
+      -- entre dos unknown (a diferencia de '||', que sí se resuelve a text, por eso
+      -- el ($2 || ' days') de abajo funciona). Escribir los números a mano en psql
+      -- no lo detecta —ahí son literales tipados—; sólo falla vía driver.
+      LIMIT GREATEST($1::int - $3::int, 1)
     ),
     adjudicaciones AS (
       SELECT o.id, o.external_code, o.status_code, o.closing_at, 3 AS priority, o.updated_at AS orden
